Create backtracking 1
Given a two dimensional array, say nxn, write an algorithm to find the total number of paths that are possible to travel from the top-left cell to bottom right cell.  You are allowed to move only right or downwards. For Ex: if nxn is 2x2, then the number of possible path is only two as shown in figure below. \

Sample Input:
n = 2
Combo of output to be printed = 3
Sample output:
1 2 4
1 3 4
(Note: The combo of output to be printed refers the number of cells that your output should hold)
For example:
for n = 3; if output to be printed is 5 then the answers will be 
1 2 3 6 9
1 2 3 6 9
and so on